Adicionada tabela pivo para relacionar produtos ao estoque, e adicionada constulta de itens por estoque

+++ app/Http/Controllers/StockController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreStockRequest;
@@ -31,7 +30,18 @@ class StockController extends Controller
 
     public function destroy(Stock $stock)
     {
+        if($stock->id == 1){
+            return redirect()->route('stock.index')->with('error', 'Não é possível excluir o grupo de armazenamento default.');
+        }
+
         $stock->delete();
         return redirect()->route('stock.index');
+    }
+
+    public function items(Stock $stock)
+    {
+        $ingredients = $stock->ingredients()->withPivot('quantity')->get();
+
+        return view('stocks.items', ['stock' => $stock, 'ingredients' => $ingredients]);
     }
 }
